Decouple router from the app
