+++ src/app.ts
@@ -1,18 +1,27 @@
 import cors from '@koa/cors';
-import Koa from 'koa';
+import Router, { RouterContext } from '@koa/router';
+import Koa, { DefaultState, Middleware } from 'koa';
 import bodyParser from 'koa-bodyparser';
 import logger from 'koa-logger';
 import Articles from './articles';
 import apiDocumentation from './middleware/api-documentation';
 import errorHandler from './middleware/error-handler';
 import routing from './middleware/routing';
-import createRouter from './router';
-import Routes from './routes';
-import { App, AppContext, AppState } from './types';
 
-export default (articles: Articles): App => {
+export type AppState = DefaultState;
+
+export type AppContext = RouterContext<AppState, {
+  articles: Articles;
+}>;
+
+export type AppMiddleware = Middleware<AppState, AppContext>;
+
+export default (
+  articles: Articles,
+  router: Router<AppState, AppContext>,
+  apiDocumentationPath: string,
+): Koa<AppState, AppContext> => {
   const app = new Koa<AppState, AppContext>();
-  const router = createRouter();
 
   app.context.articles = articles;
   app.context.router = router;
@@ -26,7 +35,7 @@ export default (articles: Articles): App => {
   app.use(cors({
     exposeHeaders: ['Link'],
   }));
-  app.use(apiDocumentation(router.url(Routes.ApiDocumentation)));
+  app.use(apiDocumentation(apiDocumentationPath));
   app.use(errorHandler());
   app.use(routing(router));
