+++ src/routes/api-documentation.ts
@@ -3,7 +3,7 @@ import { Next } from 'koa';
 import {
   hydra, owl, rdf, schema,
 } from 'rdf-namespaces';
-import { AppContext, AppMiddleware } from '../types';
+import { AppContext, AppMiddleware } from '../app';
 import Routes from './index';
 
 export default (): AppMiddleware => (
