+++ test/app.test.ts
@@ -3,14 +3,17 @@ import parseLinkHeader from 'parse-link-header';
 import request from 'supertest';
 import InMemoryArticles from '../src/adaptors/in-memory-articles';
 import createApp from '../src/app';
+import createRouter from '../src/router';
+import Routes from '../src/routes';
 
 const trim = (value: string): string => value.trim();
 const parseHeader = (header: string): Array<string> => header.split(',').map(trim);
 
+const router = createRouter();
 let app: Koa;
 
 beforeEach((): void => {
-  app = createApp(new InMemoryArticles());
+  app = createApp(new InMemoryArticles(), router, router.url(Routes.ApiDocumentation));
 });
 
 describe('the application', (): void => {
